chore: add ref forwarding to feature HOC
